backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bidict==0.23.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
defusedxml==0.7.1
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
oauthlib==3.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-decouple==3.8
python-engineio==4.12.2
python-socketio==5.13.0
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.9.0
simple-websocket==1.1.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wsproto==1.2.0
zope.interface==7.2
